+++ changelog.md
@@ -1,2 +1,3 @@
 # 0.1.3
-- Fixed no casting recipes 
+- Fixed no Controller recipe
+  - Why am I like this?